Designate recipient: Document that a holder must exist
We have to use any_realm_title to locate the title to assign to, rather than
any_title, since using any_title causes it to crash.

Unfortunately any_realm_title doesn't scope to unheld titles, so we can't
assign to titles that don't exist.

Rather than this being hidden by the fact that we use holder_scope and don't
have any provision for the title not being held, let's add an extra conditional
to *force* this and document it, so that it's obvious.