Prepare for generating more variables
I want to rework the title creation logic to store a reference to the newly
created Duchy/Kingdom/Empire in the County/Duchy/Kingdom,
so that if the title is destroyed and un-createable,
it can still be retrieved.

To do this would require a reference to be kept in the title,

but initially I'm just going to use it to replace the flag for whether there
has been a new title created from it,
and to find the newly created title when assigning de-jure territory.